[FEAT] 음식점 추가 기능 구현

+++ src/main/java/com/example/deliverybackend/application/member/GeometryPoint.java
@@ -7,7 +7,7 @@ import org.locationtech.jts.geom.Point;
 import org.locationtech.jts.geom.PrecisionModel;
 
 public class GeometryPoint {
-    public static Point createPoint(Long latitude, Long longitude) {
+    public static Point createPoint(Double latitude, Double longitude) {
         GeometryFactory geometryFactory = new GeometryFactory(new PrecisionModel(),4326);
         return geometryFactory.createPoint(new Coordinate(latitude, longitude));
     }
